+++ resolwe_bio/management/commands/generate_diffexpr.py
@@ -7,6 +7,7 @@ Generate Differential Expressions
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import csv
 import gzip
 import json
 import logging
@@ -20,7 +21,7 @@ from django.core.management.base import BaseCommand
 from django.conf import settings
 from django.utils import timezone
 
-from resolwe.flow.models import Data
+from resolwe.flow.models import Data, Storage
 from resolwe.utils import BraceMessage as __
 from resolwe_bio.models import Sample
 from .utils import get_descriptorschema, get_process, get_superuser, generate_sample_desciptor
@@ -48,6 +49,7 @@ class Command(BaseCommand):
                             help="Number of differential expressions to generate (default: %(default)s)")
         parser.add_argument('-g', '--group-size', type=int, default=5,
                             help="Number of samples in case/control group per DE (default: %(default)s)")
+        parser.add_argument('--rseed', action='store_true', help="Use fixed random seed")
 
     @staticmethod
     def get_random_word(length):
@@ -72,13 +74,14 @@ class Command(BaseCommand):
                 process=get_process('upload-cxb'),
                 contributor=get_superuser(),
                 status=Data.STATUS_PROCESSING,
-                input={'src': {'file': cuffquant_file}})
+                input={'src': {'file': cuffquant_file}, 'source': 'hg19'})
 
             os.mkdir(os.path.join(self.data_dir, str(exp.id)))
             shutil.copy(os.path.join(self.test_files_path, cuffquant_file), os.path.join(self.data_dir, str(exp.id)))
 
             exp.output = {
                 'cxb': {'file': cuffquant_file},
+                'source': 'hg19'
             }
             exp.status = Data.STATUS_DONE
             exp.save()
@@ -105,7 +108,7 @@ class Command(BaseCommand):
             process=get_process('upload-gtf'),
             contributor=get_superuser(),
             status=Data.STATUS_PROCESSING,
-            input={'src': {'file': filename}})
+            input={'src': {'file': filename}, 'source': 'hg19'})
 
         os.mkdir(os.path.join(self.data_dir, str(ann.id)))
 
@@ -114,7 +117,8 @@ class Command(BaseCommand):
                 shutil.copyfileobj(gzfile, outfile)
 
         ann.output = {
-            'gtf': {'file': filename[:-3]}
+            'gtf': {'file': filename[:-3]},
+            'source': 'hg19'
         }
         ann.status = Data.STATUS_DONE
         ann.save()
@@ -126,10 +130,68 @@ class Command(BaseCommand):
 
         return ann
 
+    @staticmethod
+    def generate_raw_data(gene_ids, path):
+        """Generate random DE data."""
+        de_data = {}
+        header = ['test_id', 'gene_id', 'gene', 'locus', 'sample_1',
+                  'sample_2', 'status', 'value_1', 'value_2',
+                  'log2(fold_change)', 'test_stat', 'p_value', 'q_value',
+                  'significant']
+
+        with gzip.open(gene_ids, mode='rt') as gene_ids:
+            all_genes = [line.strip() for line in gene_ids]
+            de_data['test_id'] = all_genes
+            de_data['gene_id'] = all_genes
+            de_data['gene'] = all_genes
+            de_data['locus'] = ['chr20:463337-524482' for gene in all_genes]
+            de_data['sample_1'] = ['control' for gene in all_genes]
+            de_data['sample_2'] = ['case' for gene in all_genes]
+            de_data['status'] = ['OK' for gene in all_genes]
+            de_data['value_1'] = [random.gammavariate(1, 100) for gene in all_genes]
+            de_data['value_2'] = [random.gammavariate(1, 100) for gene in all_genes]
+            de_data['log2(fold_change)'] = [random.uniform(-10, 10) for gene in all_genes]
+            de_data['test_stat'] = [random.uniform(-3, 3) for gene in all_genes]
+            de_data['p_value'] = [random.uniform(0, 1) for gene in all_genes]
+            de_data['q_value'] = [random.uniform(0, 1) for gene in all_genes]
+            de_data['significant'] = [random.choice(['yes', 'no']) for gene in all_genes]
+
+            rows = zip(de_data['test_id'], de_data['gene_id'], de_data['gene'],
+                       de_data['locus'], de_data['sample_1'], de_data['sample_2'],
+                       de_data['status'], de_data['value_1'], de_data['value_2'],
+                       de_data['log2(fold_change)'], de_data['test_stat'],
+                       de_data['p_value'], de_data['q_value'], de_data['significant'])
+
+            with gzip.open(os.path.join(path, 'de_raw.tab.gz'), 'wt') as raw_df:
+                writer = csv.writer(raw_df, delimiter=str('\t'), lineterminator='\n')
+                writer.writerow(header)
+                for row in rows:
+                    writer.writerow(row)
+
+        with open(os.path.join(path, 'de_json.json'), 'w') as json_file:
+            de_data_std = {
+                'stat': de_data['test_stat'],
+                'logfc': de_data['log2(fold_change)'],
+                'pvalue': de_data['p_value'],
+                'fdr': de_data['q_value'],
+                'gene_id': de_data['gene_id']
+            }
+            json.dump({'de_json': de_data_std}, json_file, indent=4, sort_keys=True)
+
+            rows = zip(de_data_std['gene_id'], de_data_std['logfc'], de_data_std['fdr'],
+                       de_data_std['pvalue'], de_data_std['stat'])
+
+            with gzip.open(os.path.join(path, 'de_file.tab.gz'), 'wt') as de_file:
+                writer = csv.writer(de_file, delimiter=str('\t'), lineterminator='\n')
+                writer.writerow(['gene_id', 'logfc', 'fdr', 'pvalue', 'stat'])
+                for row in rows:
+                    writer.writerow(row)
+
     def generate_diffexp_data(self, group_size):
         """Generate differential expression data."""
         de_name = 'cuffdiff'
-        de_file = os.path.join(self.test_files_path, de_name + '.tab.gz')
+
+        human_genes = os.path.join(self.test_files_path, 'human_genes.tab.gz')
 
         logger.info('---- case samples ----')
         case_samples = self.create_expressions(group_size)
@@ -144,8 +206,10 @@ class Command(BaseCommand):
         genome_annotation_input = self.create_genome_annotation('hg19_chr20_small.gtf.gz')
 
         de_descriptor = {
-            'fc_threshold': 2,
-            'fdr_threshold': 0.01,
+            'thresholds': {
+                'prob_field': 'fdr',
+                'logfc': 2,
+                'prob': 0.05},
             'case_label': 'Case group',
             'control_label': 'Control group'
         }
@@ -170,22 +234,31 @@ class Command(BaseCommand):
             contributor=get_superuser(),
             input=de_inputs)
 
-        # Create data directory and copy files into it
+        # Create data directory
         os.mkdir(os.path.join(self.data_dir, str(de_obj.id)))
-        shutil.copy(de_file, os.path.join(self.data_dir, str(de_obj.id)))
+        self.generate_raw_data(human_genes, os.path.join(self.data_dir, str(de_obj.id)))
 
-        # Get JSON for volcano plot
-        with gzip.open(os.path.join(self.test_files_path, '{}.json.gz'.format(de_name)), 'rt') as f:
-            de_data = json.load(f)
+        with open(os.path.join(self.data_dir, str(de_obj.id), 'test.txt'), 'w') as _:
+            pass
+
+        json_object = Storage.objects.create(
+            json=json.load(open(os.path.join(self.data_dir, str(de_obj.id), 'de_json.json'))),
+            contributor=get_superuser(),
+            name='{}_storage'.format(de_obj.name),
+            data=de_obj)
+
+        os.remove(os.path.join(self.data_dir, str(de_obj.id), 'de_json.json'))
 
         # TODO: reference on existing true files
         de_obj.output = {
-            'diffexp': {'file': de_file},
-            'de_data': de_data,
-            'transcript_diff_exp': {'file': de_file},
-            'cds_diff_exp': {'file': de_file},
-            'tss_group_diff_exp': {'file': de_file},
-            'cuffdiff_output': {'file': de_file}
+            'raw': {'file': 'de_raw.tab.gz'},
+            'de_json': json_object.id,
+            'de_file': {'file': 'de_file.tab.gz'},
+            'transcript_diff_exp': {'file': 'test.txt'},
+            'cds_diff_exp': {'file': 'test.txt'},
+            'tss_group_diff_exp': {'file': 'test.txt'},
+            'cuffdiff_output': {'file': 'test.txt'},
+            'source': 'hg19'
         }
 
         de_obj.status = Data.STATUS_DONE
@@ -200,5 +273,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         """Command handle."""
+        if options['rseed']:
+            random.seed(42)
         for _ in range(options['n_diffexps']):
             self.generate_diffexp_data(options['group_size'])