+++ resolwe_bio/tools/parse_diffexp.py
@@ -1,6 +1,4 @@
 #!/usr/bin/env python2
-# pylint: disable=missing-docstring,invalid-name
-# XXX: Refactor to a comand line tool and remove pylint disable
 """Parse Diff Exp output files."""
 from __future__ import absolute_import, division, print_function
 
@@ -11,23 +9,67 @@ import numpy as np  # pylint: disable=import-error
 import pandas as pd  # pylint: disable=import-error
 
 
-parser = argparse.ArgumentParser(description="Parse Diff Exp output files")
-parser.add_argument('de_file', help="DE analysis output file (.tab).")
-args = parser.parse_args()
+def parse_arguments():
+    """Parse command line arguments."""
+    parser = argparse.ArgumentParser(description="Parse Diff Exp output files")
+    parser.add_argument('raw_file', help="DE analysis output file (.tab).")
+    parser.add_argument('--gene_id', help="Gene_IDs column name", type=str)
+    parser.add_argument('--fdr', help="FDR column name", type=str)
+    parser.add_argument('--pvalue', help="Pvalue column name", type=str)
+    parser.add_argument('--fwer', help="FWER column name", type=str)
+    parser.add_argument('--logodds', help="Log Odds column name", type=str)
+    parser.add_argument('--logfc', help="logfc column name", type=str)
+    parser.add_argument('--stat', help="Statistics column name", type=str)
+    return parser.parse_args()
 
-de_data = pd.read_csv(args.de_file, sep='\t')
-de_data.rename(columns={'Unnamed: 0': 'gene_id'}, inplace=True)
-de_data.fillna(value=0, inplace=True)
-columns = {}
 
-for column in de_data:
-    # replace inf values in cuffdiff results file
-    if column == 'log2(fold_change)':
-        col = np.array(de_data[column])
+def main():
+    """Invoked when run directly as a program."""
+    args = parse_arguments()
+
+    de_data = pd.read_csv(args.raw_file, sep='\t')
+    de_data.rename(columns={'Unnamed: 0': 'gene_id'}, inplace=True)
+    de_data.fillna(value=0, inplace=True)
+    columns = {}
+    col_order = []
+
+    if args.gene_id:
+        columns['gene_id'] = list(de_data[args.gene_id])
+        col_order.append('gene_id')
+
+    if args.logfc:
+        col = np.array(de_data[args.logfc])
         col[np.isinf(col)] = 0
-        columns[column] = list(col)
-    else:
-        columns[column] = list(de_data[column])
+        columns['logfc'] = list(col)
+        col_order.append('logfc')
 
-data = {'de_data': columns}
-print(json.dumps(data, separators=(',', ':'), allow_nan=False))
+    if args.fdr:
+        columns['fdr'] = list(de_data[args.fdr])
+        col_order.append('fdr')
+
+    if args.pvalue:
+        columns['pvalue'] = list(de_data[args.pvalue])
+        col_order.append('pvalue')
+
+    if args.fwer:
+        columns['fwer'] = list(de_data[args.fwer])
+        col_order.append('fwer')
+
+    if args.logodds:
+        columns['logodds'] = list(de_data[args.logodds])
+        col_order.append('logodds')
+
+    if args.stat:
+        columns['stat'] = list(de_data[args.stat])
+        col_order.append('stat')
+
+    data = {'de_json': columns}
+    print(json.dumps(data, separators=(',', ':'), allow_nan=False))
+
+    outdf = pd.DataFrame(columns)
+    outdf = outdf[col_order]
+    outdf.to_csv('de_file.tab.gz', sep='\t', index=False, compression='gzip')
+
+
+if __name__ == "__main__":
+    main()